Create Store Database
This SQL project was for a Khan Acadamy project where I created a database for a store that had multiple items and columns. In this database, I listed movies that a movie store would carry, the amount of those movies they had in stock, their price, and when that movie was released. Then I used different select statements to show the data in a specific order as well as using an aggregate for a specific group. This project allowed me to practice basic SQL commands and understand the functions.