Create Digging Through Data
This SQL project was for a Khan Acadamy project where when presented with a database, I could use advanced queries to discover insight about the data. With the example database, I was able to practice commands such as CASE, HAVING, AND/OR, as well as aggregate commands like SUM, AVG, MIN, MAX, etc leading to a better understanding of how these commands work. I was also able to practice the order in which these commands go together.