Create Relational Queries
This SQL project was for a Khan Acadamy project where I created multiple tables surrounding the same topic: celebrities. The various tables give various specific information related to a celebrity. With these tables, I then practiced the different JOIN functions for joining separate tables together and joining a table to itself. It was interesting to put into practice this function and configure how the table represented the data I was trying to look for.