Create Altering App Data
This SQL project was for a Khan Acadamy project where I created a table for an app and what data they could store in it. Initially, I chose to portray a fictional fighting game app and the data stored would include the user's id, the character they chose, and the score they received. The purpose of this project was to practice commands that would alter the table such as ALTER, UPDATE, and DELETE. With this, I was able to understand how these commands could be used as well as how easy it can be to manipulate the data and the structure of the table.